BZ:2001552 close loop
