+++ tests/foreman/ui/test_rhcloud_insights.py
@@ -382,7 +382,7 @@ def test_insights_registration_with_capsule(
         cmd = session.host.get_register_command(
             {
                 'general.operating_system': default_os.title,
-                'general.orgnization': org.name,
+                'general.organization': org.name,
                 'general.capsule': rhcloud_capsule.hostname,
                 'general.activation_keys': ak.name,
                 'general.insecure': True,
